refactor: 소모임 최대 관심사 VO -> Long 수정
- #20

+++ src/main/java/com/example/wegather/group/dto/CreateSmallGroupRequest.java
@@ -23,5 +23,5 @@ public class CreateSmallGroupRequest {
   private Double latitude;
   @NotNull
   @Min(0)
-  private Integer maxMemberCount;
+  private Long maxMemberCount;
 }
