+++ src/main/java/com/example/wegather/group/dto/UpdateSmallGroupRequest.java
@@ -21,5 +21,5 @@ public class UpdateSmallGroupRequest {
   private Double latitude;
   @NotNull
   @Min(0)
-  private Integer maxMemberCount;
+  private Long maxMemberCount;
 }
